+++ src/editors/subscription/later-binding/ext-ref-later-binding-list.ts
@@ -14,14 +14,14 @@ import {
   cloneElement,
   compareNames,
   getDescriptionAttribute,
-  getNameAttribute,
   identity,
   newActionEvent,
   Replace,
+  getSclSchemaVersion,
 } from '../../../foundation.js';
 
-import { styles, updateExtRefElement } from '../foundation.js';
-import { FcdaSelectEvent, getFcdaTitleValue } from './foundation.js';
+import { styles, updateExtRefElement, serviceTypes } from '../foundation.js';
+import { FcdaSelectEvent } from './foundation.js';
 
 /**
  * A sub element for showing all Ext Refs from a FCDA Element.
@@ -31,9 +31,11 @@ import { FcdaSelectEvent, getFcdaTitleValue } from './foundation.js';
 export class ExtRefLaterBindingList extends LitElement {
   @property({ attribute: false })
   doc!: XMLDocument;
+  @property()
+  controlTag!: 'SampledValueControl' | 'GSEControl';
 
   @state()
-  currentSelectedSvcElement: Element | undefined;
+  currentSelectedControlElement: Element | undefined;
   @state()
   currentSelectedFcdaElement: Element | undefined;
   @state()
@@ -77,7 +79,7 @@ export class ExtRefLaterBindingList extends LitElement {
   }
 
   private async onFcdaSelectEvent(event: FcdaSelectEvent) {
-    this.currentSelectedSvcElement = event.detail.svc;
+    this.currentSelectedControlElement = event.detail.controlElement;
     this.currentSelectedFcdaElement = event.detail.fcda;
 
     // Retrieve the IED Element to which the FCDA belongs.
@@ -97,6 +99,33 @@ export class ExtRefLaterBindingList extends LitElement {
     );
   }
 
+  private checkEditionSpecificRequirements(extRefElement: Element): boolean {
+    if (getSclSchemaVersion(extRefElement.ownerDocument) === '2003')
+      return true;
+    return (
+      extRefElement.getAttribute('serviceType') ===
+        serviceTypes[this.controlTag] &&
+      extRefElement.getAttribute('srcLDInst') ===
+        this.currentSelectedControlElement
+          ?.closest('LDevice')
+          ?.getAttribute('inst') &&
+      (extRefElement.getAttribute('scrPrefix') || '') ===
+        (this.currentSelectedControlElement
+          ?.closest('LN0')
+          ?.getAttribute('prefix') || '') &&
+      extRefElement.getAttribute('srcLNClass') ===
+        this.currentSelectedControlElement
+          ?.closest('LN0')
+          ?.getAttribute('lnClass') &&
+      (extRefElement.getAttribute('srcLNInst') || '') ===
+        this.currentSelectedControlElement
+          ?.closest('LN0')
+          ?.getAttribute('inst') &&
+      extRefElement.getAttribute('srcCBName') ===
+        this.currentSelectedControlElement?.getAttribute('name')
+    );
+  }
+
   /**
    * Check if specific attributes from the ExtRef Element are the same as the ones from the FCDA Element
    * and also if the IED Name is the same. If that is the case this ExtRef subscribes to the selected FCDA
@@ -113,7 +142,8 @@ export class ExtRefLaterBindingList extends LitElement {
       this.sameAttributeValue(extRefElement, 'lnClass') &&
       this.sameAttributeValue(extRefElement, 'lnInst') &&
       this.sameAttributeValue(extRefElement, 'doName') &&
-      this.sameAttributeValue(extRefElement, 'daName')
+      this.sameAttributeValue(extRefElement, 'daName') &&
+      this.checkEditionSpecificRequirements(extRefElement)
     );
   }
 
@@ -187,7 +217,7 @@ export class ExtRefLaterBindingList extends LitElement {
     if (
       !this.currentIedElement ||
       !this.currentSelectedFcdaElement ||
-      !this.currentSelectedSvcElement!
+      !this.currentSelectedControlElement!
     ) {
       return null;
     }
@@ -197,7 +227,7 @@ export class ExtRefLaterBindingList extends LitElement {
       new: {
         element: updateExtRefElement(
           extRefElement,
-          this.currentSelectedSvcElement,
+          this.currentSelectedControlElement,
           this.currentSelectedFcdaElement
         ),
       },
@@ -205,21 +235,8 @@ export class ExtRefLaterBindingList extends LitElement {
   }
 
   private renderTitle(): TemplateResult {
-    const svcName = this.currentSelectedSvcElement
-      ? getNameAttribute(this.currentSelectedSvcElement)
-      : undefined;
-    const fcdaName = this.currentSelectedFcdaElement
-      ? getFcdaTitleValue(this.currentSelectedFcdaElement)
-      : undefined;
-
     return html`<h1>
-      ${translate(
-        'subscription.laterBinding.extRefList.SampledValueControl.title',
-        {
-          svcName: svcName ?? '-',
-          fcdaName: fcdaName ?? '-',
-        }
-      )}
+      ${translate(`subscription.laterBinding.extRefList.title`)}
     </h1>`;
   }
 
@@ -323,7 +340,7 @@ export class ExtRefLaterBindingList extends LitElement {
 
   render(): TemplateResult {
     return html` <section tabindex="0">
-      ${this.currentSelectedSvcElement && this.currentSelectedFcdaElement
+      ${this.currentSelectedControlElement && this.currentSelectedFcdaElement
         ? html`
             ${this.renderTitle()}
             <filtered-list>