+++ test/unit/editors/ied/__snapshots__/server-container.test.snap.js
@@ -1,7 +1,7 @@
 /* @web/test-runner snapshot v1 */
 export const snapshots = {};
 
-snapshots["server-container looks like the latest snapshot"] = 
+snapshots["server-container Server Element with LDevice Elements and all LN Elements of the LDevice Element displayed looks like the latest snapshot"] = 
 `<action-pane tabindex="0">
   <mwc-icon slot="icon">
   </mwc-icon>
@@ -11,5 +11,31 @@ snapshots["server-container looks like the latest snapshot"] =
   </ldevice-container>
 </action-pane>
 `;
-/* end snapshot server-container looks like the latest snapshot */
+/* end snapshot server-container Server Element with LDevice Elements and all LN Elements of the LDevice Element displayed looks like the latest snapshot */
+
+snapshots["server-container Server Element with LDevice Elements and some LN Elements displayed looks like the latest snapshot"] = 
+`<action-pane tabindex="0">
+  <mwc-icon slot="icon">
+  </mwc-icon>
+  <ldevice-container>
+  </ldevice-container>
+</action-pane>
+`;
+/* end snapshot server-container Server Element with LDevice Elements and some LN Elements displayed looks like the latest snapshot */
+
+snapshots["server-container Server Element with LDevice Elements and no LN Elements displayed looks like the latest snapshot"] = 
+`<action-pane tabindex="0">
+  <mwc-icon slot="icon">
+  </mwc-icon>
+</action-pane>
+`;
+/* end snapshot server-container Server Element with LDevice Elements and no LN Elements displayed looks like the latest snapshot */
+
+snapshots["server-container Server Element without LDevice Element looks like the latest snapshot"] = 
+`<action-pane tabindex="0">
+  <mwc-icon slot="icon">
+  </mwc-icon>
+</action-pane>
+`;
+/* end snapshot server-container Server Element without LDevice Element looks like the latest snapshot */