+++ test/unit/editors/subscription/goose/__snapshots__/goose-list.test.snap.js
@@ -4,9 +4,9 @@ export const snapshots = {};
 snapshots["goose-list looks like the latest snapshot with a document loaded"] = 
 `<section tabindex="0">
   <h1>
-    [subscription.goose.publisherGoose.title]
+    [subscription.goose.publisher.title]
   </h1>
-  <filtered-list>
+  <filtered-list activatable="">
     <mwc-list-item
       aria-disabled="false"
       graphic="icon"
@@ -132,9 +132,9 @@ snapshots["goose-list looks like the latest snapshot with a document loaded"] =
 snapshots["goose-list looks like the latest snapshot without a doc loaded"] = 
 `<section tabindex="0">
   <h1>
-    [subscription.goose.publisherGoose.title]
+    [subscription.goose.publisher.title]
   </h1>
-  <filtered-list>
+  <filtered-list activatable="">
   </filtered-list>
 </section>
 `;